+++ Molecular_Dynamics_Simulations/Rap_2_LJP.txt
@@ -1,0 +1,8 @@
+deltaT	0.005
+density	0.8
+initUcell_x	20
+initUcell_y	20
+stepAvg	100
+stepEquil	0
+stepLimit	500
+temperature	1.
