+++ README.md
@@ -1,2 +1,9 @@
+<<<<<<< Updated upstream
 # nanocomposite-fab #
 Making, modeling and simulating materials doped with nanosemiconductors as well as drawing them into optical fibers.
+=======
+# fabconcept #
+Making, modeling and simulating materials doped with nanosemiconductors as well as drawing them into optical fibers.
+
+For further information, see the wiki
+>>>>>>> Stashed changes
