--- protocols/NSC_purification.md
@@ -1,19 +1,0 @@
-<!--- UNDER CONSTRUCTION
-# Nanosemiconductor Purification #
-- Fume Hood
-- Centrifuge tubes (Eppendorf)
-- Micropipettes and suitable tips
-- Solvents
-TODO : Everything here, the concept and a full detailed protocol needs to be written. The following from Carly can inspire, but it's meant first for synthesis and polymerization: 
-
-CARLY
-cQD Synthesis and Dispersion. The cQDs prepared for this study consist of a CdSe core (diameter ∼3.2 nm) surrounded by a CdS shell, which were synthesized following the methods by Nasilowski et al.63 Afterward, the cQDs were purified using several centrifugation cycles with isopropyl alcohol and hexanes as the cQD nonsolvent and solvent, respectively, then redispersed in 10 mL of hexane to obtain a 6 μM concentration.
-
-For polymerization : A fixed volume of the cQD solution (Table 1) was further purified by precipitation and centrifugation with isopropyl alcohol then redispersed progressively in styrene and transferred to a glass test tube.
-
-ARTHUR
-En parallèle, un certain volume de cQDs dans une solution d’hexane est transvidé dans un tube de centrifugation. Ce dernier est rempli avec de l’éthanol anhydre, un mauvais solvant permettant de faire agréger les cQDs. Après centrifugation à 10 000rpm pendant 5 minutes,les cQD sont précipités au fond du tube de centrifugation. En remplissant le tube de centrifu-gation de styrène purifié, les cQD peuvent alors être récupérés et transvidés dans l’éprouvette contenant le reste du styrène. Cette purification des cQD permet à la fois de limiter la quantité de ligands en sursaturation dans la solution, mais surtout d’éviter d’avoir de l’hexane ensolution lors de la polymérisation. En effet, il a été observé que la présence d’hexane est àl’origine de bulle lors de l’étirage subséquent.
-
-
-**Resources on Precipitation and Centrifugation**
---->
