--- protocols/monomer_purification.md
@@ -1,62 +1,0 @@
-# Monomer Purification - Flash Column Chromatography #
-_This protocol has been tested so far with styrene from ...._
-
-Monomers are usually stabilized (stab.) with inhibitors preventing unintentional polymerization and should be kept refrigerated (unless it says otherwise on the bottle). A purification step is thus required to get rid of the inhibitors and enable polymerization.
-
-<!--- TODO: Add Cie providing monomer + grade above, hm moins bon dans un concept de protocole générique. P-e tourner en tableau listing of matching chemicals referred with a, b, c,... si ça s'allonge et qu'il faut modifier les particules ou autre pour que ça marche? Also confirm how much monomer a single column can purify and modify the N.B. as needed. --->
-## Chemicals & Furnitures ##
-- Slightly more than the amount of liquid monomer required for the polymerization
-- Pasteur pipette and its bulb
-- Small amount of glass wool and a stick to push it in the pipette
-- Funnel fitting the Pasteur pipette 
-- Basic alumina powder : Al<sub>2</sub>O<sub>3</sub> particles sieved by a 60 mesh (Alfa Aesar, Brockmann & Schodder activity grade I) 
-- Micropipettes and suitable tips
-- Small beakers
-
-## Column Packing ##
-- Plug the bottom of the Pasteur pipette with glass wool to prevent the particles from leaking out at the next step.
-- Clamp the pipette to a stand.
-- Using a funnel, fill the pipette up to its indent with the Al<sub>2</sub>O<sub>3</sub> particles.
-- To tamp down the particles by applying air pressure, put the bulb on top of the column and squeeze it, then remove it while still squeezed.
-
-_N.B. Each column can purify up to ~12 mL of styrene._  
-
-## Monomer Loading ## 
- - Place an empty beaker under the Pasteur pipette to collect the purified monomer.
- - Put the desired amount of monomer in another beaker.
- - With a micropipette, pour some monomer liquid into the column, nearly filling it.
- - To accelerate the flow while making sure sure the monomer level does not go below the top of the particles, place the bulb on the pipette and squeeze it, then remove it while still squeezed.
- - Alternatively, let gravity do the work while you do something alse, but keeping an eye on the particles to avoid drying them out.
- - Add monomer to the column until it is all purified. 
- 
-## Column Disposal ## 
-- Place a container below the Pasteur pipette to transfer waste afterwards in the proper disposal bin.
-- Load the column with hexane then acetone to flush out remaining monomers, again squeezing a bulb on top to accelerate the solvent flow. 
-- Empty the waste container into the white plastic jug for nonhalogenated solvents.
-- With the container back under the pipette, turn the latter upside down in the clamp.
-- Let this dry overnight in a fumehood until the column is emptied of particles, gently tapping it at the end if necessary.
-- To speed up drying, air or nitrogen gas(too costly? check with Stéphane) can be flowed through the column.
-- Dispose of the alumina particles in the white trash can dedicated to them.
-- The Pasteur pipette can be reused for the same monomer or put in the glassware waste bin after removing the glass wool.
- 
-<!--- TODO: Eventually try the slurry method for bigger volumes ---> 
-
-**Resources on Column Chromatography**
-<!---
-TODO : format properly as references, checking my google docs CCV & revtex to settle on a style
---->
-- <a href="https://www.orgchemboulder.com/Technique/Procedures/Columnchrom/Procedure.shtml">https://www.orgchemboulder.com/Technique/Procedures/Columnchrom/Procedure.shtml</a> [English]
-- <a href="https://www.orgchemboulder.com/Labs/Experiments/8%20-%20Column%20Chromatography.pdf">https://www.orgchemboulder.com/Labs/Experiments/8%20-%20Column%20Chromatography.pdf</a> [English]
-- <a href="https://www.orgchemboulder.com/Technique/Procedures/Columnchrom/Columnchrom.shtml">https://www.orgchemboulder.com/Technique/Procedures/Columnchrom/Columnchrom.shtml</a> [English]
-- <a href="https://doi.org/10.1002/chemv.201200074">https://doi.org/10.1002/chemv.201200074</a> [English]
-- <a href="https://chem.libretexts.org/Bookshelves/Organic_Chemistry/Organic_Chemistry_Lab_Techniques_(Nichols)/02%3A_Chromatography/2.04%3A_Column_Chromatography/2.4B%3A_Microscale_(Pipette)_Columns">https://chem.libretexts.org/Bookshelves/Organic_Chemistry/Organic_Chemistry_Lab_Techniques_(Nichols)/02%3A_Chromatography/2.04%3A_Column_Chromatography/2.4B%3A_Microscale_(Pipette)_Columns</a> [English]
-- <a href="https://www.chemistryviews.org/details/education/2040151/Tips_and_Tricks_for_the_Lab_Column_Packing/">https://www.chemistryviews.org/details/education/2040151/Tips_and_Tricks_for_the_Lab_Column_Packing/</a> [English]
-- section 2.2 Chromatographie préparative de <a href="https://doi.org/10.51257/a-v2-p1445">https://doi.org/10.51257/a-v2-p1445</a> [French]
-- <a href="https://www.silicycle.com/ca/media/pdf/applications/appn-sf002-0-brockmann-schodder-activity-test.pdf">https://www.silicycle.com/ca/media/pdf/applications/appn-sf002-0-brockmann-schodder-activity-test.pdf</a> [English]
-<!--- TODO : find a good video from the channels on Quick Notes--->
-
-
-## Resources on Other Separation Techniques ##
-- <a href="https://en.wikipedia.org/wiki/List_of_purification_methods_in_chemistry">https://en.wikipedia.org/wiki/List_of_purification_methods_in_chemistry</a> [English]
-- <a href="https://fac.umc.edu.dz/inataa/assets/files/Cours-en-Ligne/L2SA/Chimie_analytique_techniques_de_separation.pdf"> https://fac.umc.edu.dz/inataa/assets/files/Cours-en-Ligne/L2SA/Chimie_analytique_techniques_de_separation.pdf</a> [French] <!--- TODO : find a better one --->
-- Distillation : <a href="https://www.techniques-ingenieur.fr/base-documentaire/procedes-chimie-bio-agro-th2/operations-unitaires-separation-gaz-liquide-42324210/distillation-absorption-j2610/">https://www.techniques-ingenieur.fr/base-documentaire/procedes-chimie-bio-agro-th2/operations-unitaires-separation-gaz-liquide-42324210/distillation-absorption-j2610/</a> [French]
